AI_CROP_DISESE_DETECTION/requirements.txt
anthropic>=0.40.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
httpx>=0.27.0
python-dotenv>=1.0.1
pydantic>=2.7.0
groq>=0.9.0
redis>=5.0.0
slowapi>=0.1.9
